feat(metrics): Support paginating samples via random seed
To give the illusion of a consistent paging, we take the hash of the limit and
offset as the seed to `groupArraySample`.
