+++ src/sentry/search/events/datasets/discover.py
@@ -1007,9 +1007,7 @@ class DiscoverDatasetConfig(DatasetConfig):
                 SnQLFunction(
                     "example",
                     required_args=[NumericColumn("column")],
-                    snql_aggregate=lambda args, alias: function_aliases.resolve_random_sample(
-                        ["timestamp", "span_id", args["column"].name], alias
-                    ),
+                    snql_aggregate=self._resolve_random_sample,
                     private=True,
                 ),
                 SnQLFunction(
@@ -1811,6 +1809,20 @@ class DiscoverDatasetConfig(DatasetConfig):
                 )
             ],
             alias,
+        )
+
+    def _resolve_random_sample(
+        self,
+        args: Mapping[str, str | Column | SelectType | int | float],
+        alias: str,
+    ) -> SelectType:
+        offset = 0 if self.builder.offset is None else self.builder.offset.offset
+        limit = 0 if self.builder.limit is None else self.builder.limit.limit
+        return function_aliases.resolve_random_sample(
+            ["timestamp", "span_id", args["column"].name],
+            alias,
+            offset,
+            limit,
         )
 
     # Query Filters
